refactor(#4): Extract ProjectScreenshotCard as reusable component
- Move RepoCard from dashboard page into components/ProjectScreenshotCard.tsx
- Export ProjectData interface for type reusability
- Add demo_url support for click-through on screenshots
- Update dashboard to use new component
- Reduce dashboard page complexity by 300+ lines
